requirements.txt
python-telegram-bot==21.4
openpyxl==3.1.5
python-dotenv==1.0.1
watchdog==4.0.1
